Compare_result is also tested with files online
