+++ tests/optimal_interpolation/test_integration.py
@@ -3,6 +3,7 @@ import datetime as dt
 import numpy as np
 
 import xhydro.optimal_interpolation.cross_validation as cv
+import xhydro.optimal_interpolation.compare_result as cr
 from xhydro.optimal_interpolation.functions.testdata import get_file
 
 
@@ -29,6 +30,9 @@ class Test_optimal_interpolation_integration:
     )
     flow_sim_info_file = get_file(
         name=dataf + "A20_HYDREP_TEST.nc", github_url=git_url, branch=branch
+    )
+    flow_l1o_info_file = get_file(
+        name=dataf + "A20_ANALYS_FLOWJ_RESULTS_CROSS_VALIDATION_L1O_TEST.nc", github_url=git_url, branch=branch
     )
 
     # Make a list with these files paths, required for the code.
@@ -130,3 +134,15 @@ class Test_optimal_interpolation_integration:
         assert len(result_flows[0]) == (self.end_date - self.start_date).days
         assert len(result_flows[1]) == (self.end_date - self.start_date).days
         assert len(result_flows[2]) == (self.end_date - self.start_date).days
+
+    def test_compare_result_compare(self):
+        files = [
+        self.selected_station_file,
+        self.corresponding_station_file,
+        self.flow_obs_info_file,
+        self.flow_sim_info_file,
+        self.flow_l1o_info_file,
+    ]
+
+        cr.compare(self.start_date, self.end_date, files)
+
